seperated logic make modulized?
