+++ db.json
@@ -6,21 +6,32 @@
       "name": "Jeff",
       "city": " IRVINE",
       "email": "jeff777@abc.com",
-      "password": 111111
+      "password": 111111,
+      "password2": 111111
     },
     {
       "id": 2,
       "name": "Jinho",
       "city": " LOS ANGELES",
       "email": "jinho777@abc.com",
-      "password": 222222
+      "password": 222222,
+      "password2": 222222
     },
     {
       "id": 3,
       "name": "Lawrence",
       "city": " SAN JOSE",
       "email": "lawrence777@abc.com",
-      "password": 333333
+      "password": 333333,
+      "password2": 333333
+    },
+    {
+      "id": 4,
+      "name": "Tim",
+      "city": "costa mesa",
+      "email": "tim@tim.com",
+      "password": 111,
+      "password2": 111
     }
   ]
 }
